Had to rename the main file for publishing
