+++ test/api_resource_tests.dart
@@ -2,7 +2,7 @@ library api_resource_tests;
 
 import 'package:test/test.dart';
 
-import '../lib/stripe.dart';
+import '../lib/stripe_dart_flutter.dart';
 
 class TestResource extends ApiResource {
   final String object = 'test';
