+++ test/api_resources/transfer_tests.dart
@@ -4,7 +4,7 @@ import 'dart:convert';
 
 import 'package:test/test.dart';
 
-import '../../lib/stripe.dart';
+import '../../lib/stripe_dart_flutter.dart';
 import '../utils.dart' as utils;
 
 import 'card_tests.dart' as card;
