+++ test/service_tests.dart
@@ -1,8 +1,7 @@
 library service_tests;
 
 import 'package:test/test.dart';
-
-import '../lib/stripe.dart';
+import '../lib/stripe_dart_flutter.dart';
 
 main() {
   group('', () {
